fix(suite): switch device shows on transport unavailable with view only
