+++ packages/suite/src/components/suite/ConnectDevicePrompt.tsx
@@ -2,7 +2,6 @@ import styled, { useTheme } from 'styled-components';
 
 import {
     variables,
-    Button,
     motionEasing,
     LottieAnimation,
     useElevation,
@@ -11,8 +10,7 @@ import {
 } from '@trezor/components';
 import { isDesktop } from '@trezor/env-utils';
 import { Translation } from 'src/components/suite';
-import { useDevice, useDispatch } from 'src/hooks/suite';
-import { goto } from 'src/actions/suite/routerActions';
+import { useDevice } from 'src/hooks/suite';
 import type { PrerequisiteType } from 'src/types/suite';
 import { motion } from 'framer-motion';
 import { Elevation, mapElevationToBackground, mapElevationToBorder } from '@trezor/theme';
@@ -28,7 +26,6 @@ const Wrapper = styled(motion.div)<{ $elevation: Elevation }>`
     border: 1px solid ${mapElevationToBorder};
     align-items: center;
     box-shadow: ${({ theme }) => theme.boxShadowElevated};
-    margin-bottom: 60px;
 `;
 
 const ImageWrapper = styled.div`
@@ -131,14 +128,8 @@ export const ConnectDevicePrompt = ({
     prerequisite,
     connected,
     showWarning,
-    allowSwitchDevice,
 }: ConnectDevicePromptProps) => {
-    const dispatch = useDispatch();
-
     const { elevation } = useElevation();
-
-    const handleSwitchDeviceClick = () =>
-        dispatch(goto('suite-switch-device', { params: { cancelable: true } }));
 
     return (
         <Wrapper
@@ -153,12 +144,6 @@ export const ConnectDevicePrompt = ({
 
                 <Text>
                     <Translation id={getMessageId({ connected, showWarning, prerequisite })} />
-
-                    {allowSwitchDevice && (
-                        <Button variant="tertiary" onClick={handleSwitchDeviceClick}>
-                            <Translation id="TR_SWITCH_DEVICE" />
-                        </Button>
-                    )}
                 </Text>
             </ElevationUp>
         </Wrapper>
